feat(ConfigInterface): ConfigContract rename ConfigInterface
