+++ src/Clients/Event/EventConfig.php
@@ -1,20 +1,19 @@
 <?php declare(strict_types=1);
 
-namespace StrictPhp\HttpClients\Clients\Store;
+namespace StrictPhp\HttpClients\Clients\Event;
 
-use StrictPhp\HttpClients\Contracts\ConfigContract;
+use StrictPhp\HttpClients\Contracts\ConfigInterface;
 use StrictPhp\HttpClients\Entities\AbstractConfig;
 
 final class Config extends AbstractConfig
 {
     public function __construct(
         public bool $enabled = true,
-        public bool $serialized = true,
     )
     {
     }
 
-    public function initFromDefaultConfig(ConfigContract $object): void
+    public function initFromDefaultConfig(ConfigInterface $object): void
     {
     }
 }
