+++ src/Clients/Store/StoreConfig.php
@@ -1,19 +1,20 @@
 <?php declare(strict_types=1);
 
-namespace StrictPhp\HttpClients\Clients\Event;
+namespace StrictPhp\HttpClients\Clients\Store;
 
-use StrictPhp\HttpClients\Contracts\ConfigContract;
+use StrictPhp\HttpClients\Contracts\ConfigInterface;
 use StrictPhp\HttpClients\Entities\AbstractConfig;
 
 final class Config extends AbstractConfig
 {
     public function __construct(
         public bool $enabled = true,
+        public bool $serialized = true,
     )
     {
     }
 
-    public function initFromDefaultConfig(ConfigContract $object): void
+    public function initFromDefaultConfig(ConfigInterface $object): void
     {
     }
 }
